+++ ChangeLog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- jackson v2.9.10
+
 ## [6.4.0] - 2019-10-09
 
 ### Added